requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
appnope==0.1.4
assemblyai==0.35.1
asttokens==3.0.0
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.11
fastjsonschema==2.21.1
filelock==3.16.1
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.16
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.1
groq==0.13.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.5
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.12.3
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
langchain==0.3.10
langchain-core==0.3.43
langchain-google-genai==2.0.11
langchain-openai==0.2.12
langchain-text-splitters==0.3.2
langgraph==0.2.56
langgraph-checkpoint==2.0.8
langgraph-sdk==0.1.43
langsmith==0.1.147
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
narwhals==1.17.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
openai==1.57.1
orjson==3.10.12
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pipreqs==0.5.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.29.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
sentence-transformers==3.3.1
setuptools==75.1.0
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.46.1
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==14.1
wheel==0.44.0
yarg==0.1.9
yarl==1.18.3
zipp==3.21.0
